+++ src/extension.ts
@@ -4,13 +4,14 @@ import * as vscode from 'vscode';
 
 import {OutlineProvider} from './outline';
 
+import { addDepth, reduceDepth } from './commands';
+
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
 	
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "outline-map" is now active!');
 
 	// // // The command has been defined in the package.json file
 	// // // Now provide the implementation of the command with registerCommand
@@ -26,7 +27,10 @@ export function activate(context: vscode.ExtensionContext) {
 	
 	let outlineProvider = new OutlineProvider(context);
 	context.subscriptions.push(
-		vscode.window.registerWebviewViewProvider(outlineProvider.viewType, outlineProvider));
+		vscode.window.registerWebviewViewProvider(outlineProvider.viewType, outlineProvider),
+		vscode.commands.registerCommand('outline-map.reduceDepth', reduceDepth.bind(null, outlineProvider)),
+		vscode.commands.registerCommand('outline-map.addDepth', addDepth.bind(null, outlineProvider)),
+	);
 }
 console.log('Congratulations, your extension "outline-map" is now active!');